Set CLI-only flags in makefile (WIP)
These flags need to be set in order to skip compilation of the
elektroid GUI application and instead only build the elektroid-cli
application. It is highly likely that I didn't do this right.

Furthermore, I changed the minimum ALSA requirement from 1.1.8
to 1.1.3 which is highly likely to not work at all. The Raspbian
system (monome norns) targeted may not be able to use ALSA 1.1.8
and therefore I wanted to try it with what was on hand before
going further.